refactor: move per-page components to each route
