+++ web-client/src/app/(home)/components/JournalList.tsx
